Added bin to .gitignore
I hope this doesn’t mess things up.
I’m doing this because the bin directory is recreated by eclipse every
time you run the code. If this is included every time you commit code
the bin will be shown to be different, even though the data in it will
never be used twice.